Репозиторий для рабочего стола

+++ server/src/main/java/com/acumendev/climatelogger/repository/dbo/DashboardItemDbo.java
@@ -3,10 +3,10 @@ package com.acumendev.climatelogger.repository.dbo;
 public class DashboardItemDbo {
 
     public final long dashboardId;
-    public final long sensorId;
+    public final int sensorId;
     public final String data;
 
-    public DashboardItemDbo(long dashboardId, long sensorId, String data) {
+    public DashboardItemDbo(long dashboardId, int sensorId, String data) {
         this.dashboardId = dashboardId;
         this.sensorId = sensorId;
         this.data = data;
